requirements.txt
streamlit
ollama
exa-py
Pillow
